exercise 4.20 - 4.21

+++ part4/controllers/blogs.js
@@ -3,15 +3,6 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-const getTokenFrom = (request) => {
-  const authorization = request.get('authorization')
-
-  if(authorization && authorization.toLowerCase().startsWith('bearer ')){
-    return authorization.substring(7)
-  }
-    return null
-}
-
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
@@ -19,11 +10,11 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response, next) => {
   const reqObject = request.body
-  const token = getTokenFrom(request)
 
   try {
-    const decodedToken = jwt.verify(token, process.env.SECRET)
-    if(!token || !decodedToken.id){
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if(!request.token || !decodedToken.id){
       return response.status(401).json({error: 'token missing or invalid'})
     }
 
@@ -47,6 +38,22 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try{
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if(!request.token || !decodedToken.id){
+      return response
+        .status(401)
+        .json({error: 'token missing or invalid'})
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if(user._id.toString() !== blog.user.toString()){
+      return response
+        .status(401)
+        .json({ error: 'invalid user' })
+    }
+    
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
   } catch(error){
